DPA-107 TopBar component add
