+++ src/App.js
@@ -3,8 +3,9 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
 // components
 import LoginPage from './Components/Authorization/LoginPage';
-import HomePage from './Components/home/HomePage';
+import HomePage from './Components/Home/HomePage';
 import PrintersPage from './Components/Printers/PrintersPage';
+import FilamentsPage from './Components/Filaments/FilamentsPage';
 
 // UI elements
 import './App.css';
@@ -27,6 +28,7 @@ function App() {
     { path: endpoints.loginPage, element: <LoginPage api={endpoints} /> },
     { path: endpoints.home, element: <HomePage api={endpoints} /> },
     { path: endpoints.printersPage, element: <PrintersPage api={endpoints} /> },
+    { path: endpoints.filamentsPage, element: <FilamentsPage api={endpoints} /> },
   ]);
 
   return <RouterProvider router={router} />;
