+++ src/Components/Filaments/FilamentsPage.js
@@ -1,0 +1,25 @@
+// libs
+import React from 'react';
+
+// components
+import TopBar from '../_shared/TopBar';
+import Button from '../UI/shared/buttons/Button';
+
+// UI elements
+import StyledLink from '../UI/shared/StyledLink';
+
+function FilamentsPage(props) {
+  return (
+    <div>
+      {/* <header> */}
+      <TopBar />
+      {/* </ header> */}
+
+      <main className='App-header'>
+        <nav></nav>
+      </main>
+    </div>
+  );
+}
+
+export default FilamentsPage;
