+++ src/Components/Home/Quotes.js
