+++ src/Components/Home/UI/_quotes.scss
