+++ src/Components/Printers/PrintersPage.js
@@ -2,11 +2,12 @@
 import React from 'react';
 
 // components
+import TopBar from '../_shared/TopBar';
 import Button from '../UI/shared/buttons/Button';
 
 // UI elements
 import logo from '../../Images/icon-white.png';
-// import StyledLink from '../UI/shared/StyledLink';
+import StyledLink from '../UI/shared/StyledLink';
 
 // scss
 import '../UI/shared/_topbar.scss';
@@ -14,81 +15,9 @@ import '../UI/shared/_topbar.scss';
 function PrintersPage(props) {
   return (
     <div>
-      <header className='topbar'>
-        <div className='topbar_logo-wrapper'>
-          <div className='logo-wrapper_txt '>FraGor StartUp</div>
-          <img
-            src={logo}
-            alt='Logo'
-          />
-        </div>
-        <div className='topbar_menu'>
-          <button>
-            <svg
-              xmlns='http://www.w3.org/2000/svg'
-              width='40'
-              height='40'
-              viewBox='0 0 40 40'
-              fill='none'
-            >
-              <circle
-                cx='35'
-                cy='5'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='20'
-                cy='5'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='5'
-                cy='5'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='35'
-                cy='20'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='20'
-                cy='20'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='5'
-                cy='20'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='35'
-                cy='35'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='20'
-                cy='35'
-                r='5'
-                fill='#D9D9D9'
-              />
-              <circle
-                cx='5'
-                cy='35'
-                r='5'
-                fill='#D9D9D9'
-              />
-            </svg>
-          </button>
-        </div>
-      </header>
+      {/* <header> */}
+      <TopBar />
+      {/* </ header> */}
 
       <main className='App-header'>
         <Button
@@ -97,6 +26,21 @@ function PrintersPage(props) {
         >
           Add Printer
         </Button>
+
+        <Button
+          className=''
+          color='blue'
+        >
+          Test printer btn
+        </Button>
+        <StyledLink to={props.api.home}>
+          <Button
+            className=''
+            color='red'
+          >
+            Back
+          </Button>
+        </StyledLink>
       </main>
     </div>
   );
